Update: output more useful trimming alingment log

+++ utils/Alignment_with_trimming.sh
@@ -9,35 +9,34 @@ DIR_LIB=$(dirname $0)
 PROGRAM_CIGAR=${DIR_LIB}/CigarFilter.pl
 cd ${DIR_DATA}
 
-echo "#=========================="
-echo "# alignment of ${FILE_fastq}"
-echo "#=========================="
-date
-echo
+FILE_log=${OUT}_bowtie2.log
+FILE_tmp=${OUT}_bowtie2_tmp.log
+echo "Length Total % NoAlign % Unique % Multiple %" | tr ' ' '\t' > $FILE_log
+function getLog(){
+	# cat $FILE_tmp
+	cat $FILE_tmp | tr -d '()' | awk -v OFS='\t' 'NR>1&&NR<6{print $1,$2}' | xargs | tr ' ' '\t' >> $FILE_log && rm $FILE_tmp
+}
+
 FILE_first=$(echo ${FILE_fastq} | cut -f1 -d',')
 let READ_LENGTH=$(head -n 2 ${FILE_first} | tail -n 1 | wc -m)
 let MAX2_TRIM=$READ_LENGTH-25
 let MAX_TRIM=$READ_LENGTH-20
 
-echo "alignment of ${READ_LENGTH}bp read" 
-bowtie2 -x ${BOWTIE2_INDEX} -U ${FILE_fastq} -q -p 12 --no-unal --un ${OUT}_unaligned.fastq > ${OUT}.sam
+echo -n "${READ_LENGTH}bp	" >> $FILE_log
+bowtie2 -x ${BOWTIE2_INDEX} -U ${FILE_fastq} -q -p 12 --no-unal --un ${OUT}_unaligned.fastq > ${OUT}.sam 2> $FILE_tmp && getLog
 mv ${OUT}_unaligned.fastq ${OUT}_tmp.fastq
-echo
 
 for LEN in $(seq 5 5 $MAX2_TRIM)
 do
-	echo "trimming $LEN bp" 
-	bowtie2 -x ${BOWTIE2_INDEX} -U ${OUT}_tmp.fastq -q -3 $LEN -p 12 --no-hd --no-unal --un ${OUT}_unaligned.fastq >> ${OUT}.sam
+	echo -n "Trim ${LEN}bp	" >> $FILE_log
+	bowtie2 -x ${BOWTIE2_INDEX} -U ${OUT}_tmp.fastq -q -3 $LEN -p 12 --no-hd --no-unal --un ${OUT}_unaligned.fastq >> ${OUT}.sam 2> $FILE_tmp && getLog
 	mv ${OUT}_unaligned.fastq ${OUT}_tmp.fastq
-	echo
 done
 
-echo "trimming $MAX_TRIM bp" 
-bowtie2 -x ${BOWTIE2_INDEX} -U ${OUT}_tmp.fastq -q -3 ${MAX_TRIM} -p 12 --no-hd  >> ${OUT}.sam
+echo -n "Trim ${MAX_TRIM}bp	" >> $FILE_log
+bowtie2 -x ${BOWTIE2_INDEX} -U ${OUT}_tmp.fastq -q -3 ${MAX_TRIM} -p 12 --no-hd  >> ${OUT}.sam 2> $FILE_tmp && getLog
 rm ${OUT}_tmp.fastq
-echo
 
-echo "alignment of partial alignment reads" 
 perl ${PROGRAM_CIGAR} -s ${OUT}.sam -o ${OUT}_fastqList.txt > ${OUT}_tmp.sam
 mv ${OUT}_tmp.sam ${OUT}.sam
 LIST_NUM=$(cat ${OUT}_fastqList.txt | wc -l)
@@ -45,23 +44,24 @@ for i in $(seq 1 ${LIST_NUM})
 do
 	TRIM_OPTION=$(head -n $i ${OUT}_fastqList.txt | tail -n 1 | cut -f1)
 	FASTQ_REANALYZE=$(head -n $i ${OUT}_fastqList.txt | tail -n 1 | cut -f2)
-	echo "Command : bowtie2 -x ${BOWTIE2_INDEX} -U ${FASTQ_REANALYZE} -q ${TRIM_OPTION} -p 12 --no-hd" 
-	bowtie2 -x ${BOWTIE2_INDEX} -U ${FASTQ_REANALYZE} -q ${TRIM_OPTION} -p 12 --no-hd >> ${OUT}.sam
+	# echo "Command : bowtie2 -x ${BOWTIE2_INDEX} -U ${FASTQ_REANALYZE} -q ${TRIM_OPTION} -p 12 --no-hd"
+	echo -n "Option:${TRIM_OPTION}	" >> $FILE_log
+	bowtie2 -x ${BOWTIE2_INDEX} -U ${FASTQ_REANALYZE} -q ${TRIM_OPTION} -p 12 --no-hd >> ${OUT}.sam 2> $FILE_tmp && getLog
 	rm ${FASTQ_REANALYZE}
-	echo
 done
 rm ${OUT}_fastqList.txt
 
-echo "convert to bam file" 
 samtools view -bS ${OUT}.sam > ${OUT}.bam
-echo "sort bam file" 
+
+### read check
+TOTAL_READ=$(cat $FILE_log | awk 'NR==2{print $2}')
+TOTAL_BAM=$(samtools view -c ${OUT}.bam)
+if [ $TOTAL_READ -ne $TOTAL_BAM ]; then
+	echo "Total reads: $TOTAL_READ and Total Bam output: $TOTAL_BAM not matched"
+	exit 1
+fi
+
 samtools sort -n ${OUT}.bam -o ${OUT}_sort.bam -T tmpBamSort_${OUT}
-echo "convert again to sam file" 
 samtools view ${OUT}_sort.bam > ${OUT}.sam 
 rm ${OUT}_sort.bam ${OUT}.bam
-
-echo "finished" 
-date 
-echo
-echo
-
+exit 0
